pyproject.toml
[project]
name = "llm-learning"
version = "0.1.0"
description = "LLM Chatbot demo using FastAPI and LangGraph"
authors = [
    {name = "HoaNS"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "langchain (>=0.3.18,<0.4.0)",
    "langchain-openai (>=0.3.5,<0.4.0)",
    "langchain-community (>=0.3.17,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "black (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-ollama (>=0.2.3,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
